% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genFriedman.R
\name{genFriedman}
\alias{genFriedman}
\title{genFriedman}
\usage{
genFriedman(
  noFeatures = 10,
  noSamples = 100,
  sigma = 1,
  bins = NULL,
  seed = NULL
)
}
\arguments{
\item{noFeatures}{Integer specifying the number of features/variables to generate. Default is 10.}

\item{noSamples}{Integer specifying the number of samples to generate. Default is 100}

\item{sigma}{Numreic specifying standard deviation of the noise.}

\item{bins}{Integer specifiying the number of bins to split responce variable into. Setting a value greater than 1 turns this into a classification problem where bins determines the number of classes.}

\item{seed}{Integer specifying the random seed.}
}
\description{
Simulate data from the Friedman benchmark problem 1. Mainly used for testing.
}
\examples{
genFriedman(noFeatures = 10, noSamples = 100, sigma = 1, seed = NULL)


}
