% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_encircle.R
\docType{data}
\name{GeomEncircle}
\alias{GeomEncircle}
\title{Geom for drawing encircling polygons around groups}
\format{
An object of class \code{GeomEncircle} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomEncircle
}
\description{
This geom is used internally by \code{\link[=geom_encircle_vivi]{geom_encircle_vivi()}}, but is exported
so that users can add it directly if desired.
}
\keyword{datasets}
