% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2df.R
\name{matrix2df}
\alias{matrix2df}
\title{matrix2df}
\usage{
matrix2df(matrix)
}
\arguments{
\item{matrix}{A matrix to be converted to a data frame.}
}
\value{
A data frame of Vimp and Vint values and their index from the matrix.
}
\description{
@description Takes a matrix, such as that created by \code{vivi}, and turn it into a data frame
containing variable names, Vimp and Vint values, and the row and column index from the original
matrix.
}
\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
myMat <- vivi(fit = rF, data = aq, response = "Ozone")
myDf <- matrix2df(myMat)
myDf

}
