% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVivi.R
\name{plot.vivid}
\alias{plot.vivid}
\title{vivid: Plot the interaction/importance matrix created from \code{vivi}}
\usage{
\method{plot}{vivid}(
  mat,
  type = c("heatMap", "network"),
  title = "",
  intPal = rev(sequential_hcl(palette = "Blues 3", n = 11)),
  impPal = rev(sequential_hcl(palette = "Reds 3", n = 11)),
  fitlimsInt = NULL,
  fitlimsImp = NULL,
  angle = NULL,
  thresholdValue = 0,
  label = FALSE,
  labelNudge = 0.05,
  layout = "circle",
  cluster = F,
  clusterType = cluster_optimal,
  clusterLayout = layout_with_fr,
  ...
)
}
\arguments{
\item{mat}{A matrix of class \code{vivid}.}

\item{type}{Type of plot required. Either "heatMap" or "network".}

\item{title}{Adds title to the plot.}

\item{intPal}{A colorspace colour palette to display the interaction values.}

\item{impPal}{A colorspace colour palette to display the importance values.}

\item{fitlimsInt}{Specifies the fit range for the color map for interaction strength.}

\item{fitlimsImp}{Specifies the fit range for the color map for importance.}

\item{angle}{The angle to display the x-axis labels.}

\item{...}{Not currently implemented.}
}
\value{
A plot chosen from the type argument.
}
\description{
This function is used to create vivid-based plots. That is, either "heatMap" or "network" plots.
}
\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
myMat <- vivi(fit = rF, data = aq, response = "Ozone")
plot(myMat, type = "heatMap")
}
