% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vividHeatmap.R
\name{vividHeatMap}
\alias{vividHeatMap}
\title{vividHeatap}
\usage{
vividHeatMap(
  mat,
  title = "",
  intPal = rev(sequential_hcl(palette = "Blues 3", n = 11)),
  impPal = rev(sequential_hcl(palette = "Reds 3", n = 11)),
  fitlimsInt = NULL,
  fitlimsImp = NULL,
  angle = NULL,
  ...
)
}
\arguments{
\item{mat}{A matrix, such as that returned by vivi, of values to be plotted.}

\item{title}{Adds title to the plot.}

\item{intPal}{A colorspace colour palette to display the interaction values.}

\item{impPal}{A colorspace colour palette to display the importance values.}

\item{fitlimsInt}{Specifies the fit range for the color map for interaction strength.}

\item{fitlimsImp}{Specifies the fit range for the color map for importance.}

\item{angle}{The angle to display the x-axis labels.}

\item{...}{Not currently implemented.}
}
\value{
A heatmap plot showing variable importance on the diagonal
and variable interaction on the off-diagonal.
}
\description{
Plots a Heatmap-style display showing variable importance on the diagonal
and variable interaction on the off-diagonal.
}
\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
myMat <- vivi(fit = rF, data = aq, response = "Ozone")
plot(myMat, type = "heatMap")


# Main plot function -----------------------------------------------------------
}
