% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivi.R
\name{vividImportance}
\alias{vividImportance}
\alias{vividImportance.default}
\alias{vividImportance.ranger}
\alias{vividImportance.randomForest}
\alias{vividImportance.Learner}
\alias{vividImportance.lda}
\title{vividImportance}
\usage{
vividImportance(
  fit,
  data,
  response = NULL,
  importanceType = NULL,
  importanceMode = NULL,
  predictFun = NULL,
  ...
)

\method{vividImportance}{default}(
  fit,
  data,
  response = NULL,
  importanceType = NULL,
  importanceMode = NULL,
  predictFun = NULL
)

\method{vividImportance}{ranger}(
  fit,
  data,
  response = NULL,
  importanceType = NULL,
  importanceMode = NULL,
  predictFun = NULL
)

\method{vividImportance}{randomForest}(
  fit,
  data,
  response = NULL,
  importanceType = NULL,
  importanceMode = NULL,
  predictFun = NULL
)

\method{vividImportance}{Learner}(
  fit,
  data,
  response = NULL,
  importanceType = NULL,
  importanceMode = NULL,
  predictFun = NULL
)

\method{vividImportance}{lda}(
  fit,
  data,
  response = NULL,
  importanceType = NULL,
  importanceMode = NULL,
  predictFun = NULL
)
}
\arguments{
\item{fit}{A supervised machine learning model, which understands condvis2::CVpredict}

\item{data}{Data frame used for fit}

\item{response}{The name of the response for the fit.}

\item{importanceType}{Set to equal "agnostic" to override embedded importance measures and return agnostic importance values.}

\item{importanceMode}{Variable importance mode. One of either "\%IncMSE" or "IncNodePurity". For use with randomForest.}

\item{predictFun}{Function of (fit, data) to extract numeric predictions from fit. Uses condvis2::CVpredict by default, which works for many fit classes.}
}
\value{
A named vector of variable importance.
}
\description{
Measures variable importance for a model fit. If the model supplied does not have
and embedded variable importance measure, then an agnostic permutation method is used.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: vividImportance method

\item \code{ranger}: vividImportance method

\item \code{randomForest}: vividImportance method

\item \code{Learner}: vividImportance method

\item \code{lda}: vividImportance method
}}

\examples{
# Measuring variable importance using an agnostoc method
aq <- na.omit(airquality)
lmf <- lm(Ozone ~ ., data = aq)
vividImportance(fit = lmf, data = aq, response = "Ozone")
}
