% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivi.R
\name{vividInteraction}
\alias{vividInteraction}
\alias{vividInteraction.default}
\title{vividInteraction}
\usage{
vividInteraction(
  fit,
  data,
  response = NULL,
  class = 1,
  interactionType = NULL,
  nmax = 500,
  gridSize = 10,
  predictFun = NULL,
  ...
)

\method{vividInteraction}{default}(
  fit,
  data,
  response = NULL,
  class = 1,
  interactionType = NULL,
  nmax = 500,
  gridSize = 10,
  predictFun = NULL
)
}
\arguments{
\item{fit}{A supervised machine learning model, which understands condvis2::CVpredict}

\item{data}{Data frame used for fit.}

\item{response}{The name of the response for the fit.}

\item{class}{Category for classification, a factor level, or a number indicating which factor level.}

\item{interactionType}{passed to vividInteraction.}

\item{nmax}{Maximum number of data rows to consider.}

\item{gridSize}{The size of the grid for evaluating the predictions.}

\item{predictFun}{Function of (fit, data) to extract numeric predictions from fit. Uses condvis2::CVpredict by default, which works for many fit classes.}
}
\value{
A matrix of values, of class vivid.
}
\description{
Creates a matrix displaying variable interaction on the off-diagonal.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: vividInteraction method
}}

\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
vividInteraction(fit = rF, data = aq, response = "Ozone")
}
