% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivi.R
\name{vividReorder}
\alias{vividReorder}
\title{vividReorder}
\usage{
vividReorder(d)
}
\arguments{
\item{d}{A matrix such as that returned by vivi}

\item{newImp}{A named vector of variable importances.}

\item{reorder}{If TRUE, uses vividReorder on the matrix}
}
\value{
A reordered version of d.
}
\description{
Reorders a square matrix so that values of high importance and
interaction strength are pushed to the top left of the matrix.
}
\examples{
f <- lm(Sepal.Length ~ ., data = iris[, -5])
m <- vivi(fit = f, data = iris[, -5], response = "Sepal.Length")
corimp <- abs(cor(iris[, -5])[1, -1])
viviUpdate(m, corimp) # use correlation as importance and reorder
}
