% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviNetwork.R
\name{viviNetwork}
\alias{viviNetwork}
\title{viviNetwork}
\usage{
viviNetwork(
  mat,
  intThreshold = 0,
  intLims = NULL,
  impLims = NULL,
  intPal = rev(sequential_hcl(palette = "Blues 3", n = 100)),
  impPal = rev(sequential_hcl(palette = "Reds 3", n = 100)),
  removeNode = FALSE,
  layout = "circle",
  cluster = NULL,
  nudge = 0.5
)
}
\arguments{
\item{mat}{A matrix, such as that returned by vivi, of values to be plotted.}

\item{intThreshold}{Remove edges with weight below this value if provided.}

\item{intLims}{Specifies the fit range for the color map for interaction strength.}

\item{impLims}{Specifies the fit range for the color map for importance.}

\item{intPal}{A vector of colours to show interactions, for use with scale_fill_gradientn.}

\item{impPal}{A vector of colours to show importance, for use with scale_fill_gradientn.}

\item{removeNode}{If TRUE, then removes nodes with no connecting edges when thresholding interaction values.}

\item{layout}{Layout of the plotted graph.}

\item{cluster}{Either a vector of cluster memberships for nodes or an igraph clustering function.}

\item{nudge}{Move labels a small distance from the nodes they are labelling. Ideal values are between 0 and 1.}
}
\value{
A plot displaying interaction strength between variables on the edges and variable importance on the nodes.
}
\description{
Create a Network plot displaying variable importance
and variable interaction.
}
\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
myMat <- vivi(fit = rF, data = aq, response = "Ozone")
viviNetwork(myMat)
}
