% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.vivid.R
\name{as.data.frame.vivid}
\alias{as.data.frame.vivid}
\title{as.data.frame.vivid}
\usage{
\method{as.data.frame}{vivid}(matrix, class = "vivid")
}
\arguments{
\item{matrix}{A matrix to be converted to a data frame.}
}
\value{
A data frame of Vimp and Vint values and their index from the matrix.
}
\description{
@description Takes a matrix, such as that created by \code{vivi}, and turn it into a data frame
containing variable names, Vimp and Vint values, and the row and column index from the original
matrix.
}
\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
myMat <- vivi(fit = rF, data = aq, response = "Ozone")
myDf <- as.data.frame(myMat, class = "vivid")
myDf
}
