% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viviNetwork.R
\name{viviNetwork}
\alias{viviNetwork}
\title{viviNetwork}
\usage{
viviNetwork(
  mat,
  threshold = 0,
  intLims = NULL,
  impLims = NULL,
  intPal = rev(sequential_hcl(palette = "Blues 3", n = 11)),
  impPal = rev(sequential_hcl(palette = "Reds 3", n = 11)),
  labelNudge = 0.05,
  layout = "circle",
  cluster = NULL
)
}
\arguments{
\item{mat}{A matrix, such as that returned by vivi, of values to be plotted.}

\item{threshold}{Remove edges with weight below this value if provided.}

\item{intLims}{Specifies the fit range for the color map for interaction strength.}

\item{impLims}{Specifies the fit range for the color map for importance.}

\item{intPal}{A vector of colours to show interactions, for use with scale_fill_gradientn.}

\item{impPal}{A vector of colours to show importance, for use with scale_fill_gradientn.}

\item{labelNudge}{A value to determine the y_postioning of the variables names. A higher value will postion the label farther above the nodes.}

\item{layout}{Layout of the plotted graph.}

\item{cluster}{Either a vector of cluster memberships for nodes or an igraph clustering function.}
}
\value{
A plot displaying interaction strength between variables on the edges and variable importance on the nodes.
}
\description{
@description Create a Network  plot displaying Variable
and Variable Interaction.
}
\examples{
library(ranger)
aq <- na.omit(airquality)
rF <- ranger(Ozone ~ ., data = aq, importance = "permutation")
myMat <- vivi(fit = rF, data = aq, response = "Ozone")
viviNetwork(myMat)
}
